Changed Method to align to nSidedDie changes
Changed how the dice were handled in this class. Moved method sumArray() over to nSidedDie, renamed the dice to reflect one singular dice object used throughout the class, removed unused imports. Extracted the for loop that rolls four dice and removes the lowest, and placed that as a method in nSidedDIe