Added the Extracted methods
Three new methods to make the job of the switch statements in main file obsolete. Character Class Prompt to simply cut out the middle man in Main when it comes to assignment of the class. Similarly with RacePrompt, which also handles the specific cases for each choice in a cleaner way with no switches. Finally added a new function getCharacterHitDice to 'remove a variable and replace it with a reference'